tools/deepspan-codegen/pyproject.toml
# SPDX-License-Identifier: Apache-2.0
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepspan-codegen"
version = "0.1.0"
description = "Code generator: hwip.yaml → C/C++/Go/Python/proto artifacts"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "typer>=0.12",
    "rich>=13.0",
]

[project.scripts]
deepspan-codegen = "deepspan_codegen.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["deepspan_codegen"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
